Portfolio
Need to add all projects into netlify to make the links in my portfolio work. Was thinking to link all the projects there so people can have a look at the code. Since I haven't spent that long on the designs for the projects I was thinking to use the portfolio more as a showcase of the different projects and what the brief was for them.